backend/requirements.txt
pandas
numpy
scikit-learn
xgboost
lightgbm
fastapi
uvicorn
joblib
requests
python-multipart
sqlalchemy
passlib[bcrypt]
python-jose[cryptography]
bcrypt
pymongo
python-dotenv
